Update proposed API
